+++ lib/reconcile/astarte_generic_api.go
@@ -57,6 +57,7 @@ func EnsureAstarteGenericAPIWithCustomProbe(cr *apiv1alpha1.Astarte, api apiv1al
 		"component":         "astarte",
 		"astarte-component": component.DashedString(),
 	}
+
 	matchLabels := map[string]string{"app": deploymentName}
 
 	// Ok. Shall we deploy?
@@ -89,7 +90,7 @@ func EnsureAstarteGenericAPIWithCustomProbe(cr *apiv1alpha1.Astarte, api apiv1al
 		Strategy: getDeploymentStrategyForClusteredResource(cr, api.AstarteGenericClusteredResource, component),
 		Template: v1.PodTemplateSpec{
 			ObjectMeta: metav1.ObjectMeta{
-				Labels: labels,
+				Labels: computePodLabels(api.AstarteGenericClusteredResource, labels),
 			},
 			Spec: getAstarteGenericAPIPodSpec(deploymentName, cr, api, component, customProbe),
 		},