+++ lib/reconcile/cassandra.go
@@ -142,7 +142,7 @@ func EnsureCassandra(cr *apiv1alpha1.Astarte, c client.Client, scheme *runtime.S
 		},
 		Template: v1.PodTemplateSpec{
 			ObjectMeta: metav1.ObjectMeta{
-				Labels: labels,
+				Labels: computePodLabels(cr.Spec.Cassandra.AstarteGenericClusteredResource, labels),
 			},
 			Spec: getCassandraPodSpec(statefulSetName, dataVolumeName, cr),
 		},